imagekit: refactor code and implemented fs.Copier
