--- backend/imagekit/client/config/cloud.go
@@ -1,8 +1,0 @@
-package config
-
-// Cloud defines the cloud configuration required to connect your application to ImageKit.io.
-type Cloud struct {
-	PrivateKey  string
-	PublicKey   string
-	UrlEndpoint string
-}
