+++ backend/imagekit/client/config/configuration.go
@@ -4,15 +4,19 @@ package config
 import (
 	"fmt"
 	"net/http"
-
-	"github.com/creasty/defaults"
 )
 
 // Configuration is the main configuration struct.
 type Configuration struct {
-	API   API
-	Cloud Cloud
-	HttpClient *http.Client
+	Prefix         string
+	UploadPrefix   string
+	MetadataPrefix string
+	Timeout        int64
+	UploadTimeout  int64
+	PrivateKey     string
+	PublicKey      string
+	UrlEndpoint    string
+	HttpClient     *http.Client
 }
 
 // NewParams is a struct to define parameters to imagekit
@@ -42,18 +46,15 @@ func New(params NewParams) (*Configuration, error) {
 
 // NewFromParams returns a new Configuration instance from the provided keys and endpointUrl.
 func NewFromParams(privateKey string, publicKey string, endpointUrl string) *Configuration {
-	cloudConf := Cloud{
-		PrivateKey:  privateKey,
-		PublicKey:   publicKey,
-		UrlEndpoint: endpointUrl,
-	}
-
-	var api = API{}
-	defaults.Set(&api)
-
 	return &Configuration{
-		Cloud: cloudConf,
-		API:   api,
-		HttpClient: &http.Client{},
+		Prefix:         "https://imagekit.io/api/v2/",
+		UploadPrefix:   "https://upload.imagekit.io/api/v2/",
+		MetadataPrefix: "https://api.imagekit.io/v1/",
+		Timeout:        60,
+		UploadTimeout:  3600,
+		PrivateKey:     privateKey,
+		PublicKey:      publicKey,
+		UrlEndpoint:    endpointUrl,
+		HttpClient:     &http.Client{},
 	}
 }
